Rename cubeworld to asteroid
